+++ src/pages/index.ts
@@ -1,5 +1,5 @@
 export * from './welcome/welcome-page';
-export * from './login';
+export * from './login/login';
 export * from './boards-list';
 export * from './selected-board/selected-board';
 export * from './profile/profile';
