+++ src/pages/login/login.scss
@@ -1,0 +1,21 @@
+* {
+  margin: 0;
+  padding: 0;
+  outline: none;
+  box-sizing: border-box;
+  font-family: 'Poppins', sans-serif;
+}
+
+.container {
+  margin: auto;
+  background: #fff;
+  max-width: 30rem;
+  padding: 10px 20px 20px 20px;
+  box-shadow: 0 0 39px rgb(0 0 0 / 22%);
+  background-color: #f0f0f0;
+}
+.message {
+  margin: 15px 0 0;
+  color: #b3b3b3;
+  font-size: 14px;
+}
