+++ src/translation/by.json
@@ -5,6 +5,8 @@
   "Header board link": "Дошки",
   "Header login link": "Увайсці",
   "Header sign up link": "Зарэгістравацца",
+  "Not registered": "Не зарэгістраваны?",
+  "Registered": "Зарэгістраваны?",
   "Header edit profile link": "Профіль",
   "Header logout link": "Выйсці",
   "Welcome": "Прывітанне",
@@ -39,6 +41,7 @@
   "Name": "Імя:",
   "Login": "Логiн:",
   "UserId": "Карыстальнік:",
+  "User registered": "Карыстальнік зарэгістраваны!",
   "Password": "Пароль:",
   "Edit board": "Рэдагаваць дошку",
   "Edit button": "Рэдагаваць",
