+++ src/translation/en.json
@@ -69,6 +69,9 @@
   "Description": "Description:",
   "Owner": "Owner:",
   "Users": "Users:",
+  "User registered": "User registered!",
+  "Not registered": "Not registered?",
+  "Registered": "Registered?",
   "Select users": "Select users",
   "Cancel button": "Cancel",
   "Popconfirm column": "Are you sure you want to delete this column?",
