+++ src/translation/ru.json
@@ -4,6 +4,8 @@
   "Header create new board": "Создать доску",
   "Header board link": "Доски",
   "Header login link": "Войти",
+  "Not registered": "Не зарегистрирован?",
+  "Registered": "Зарегистрирован?",
   "Header sign up link": "Регистрация",
   "Header edit profile link": "Профиль",
   "Header logout link": "Выйти",
@@ -16,6 +18,7 @@
   "Input password": "Введите пароль",
   "Input login": "Введите логин",
   "Input name": "Введите имя",
+  "User registered": "Пользователь зарегистрирован!",
   "Welcome": "Привет",
   "Welcome title": "Добро пожаловать в Teamber",
   "Welcome text": "Teamber - новое уникальное приложение которое поможет управлять рабочим процессом более эффективно. Создавай задания, работай со своей командой и достигай целей. \"Собраться вместе - это начало. Держаться вместе - это прогресс. Работать вместе - успех.\" - Генри Форд",
